+++ src/main/java/lotto/enums/LottoInfo.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.enums;
 
 public enum LottoInfo {
     START_NUMBER(1),
